Moving error list to react lib custom form class
